Changed CLI task print method. Previous method was causing memory corruption.

+++ src/bmc_rx.h
@@ -71,7 +71,7 @@ struct bmcChannel {
 typedef struct bmcChannel bmcChannel;
 
 struct bmcChannels {
-    uint8_t maxChannels;	// Number of channels in pointer array
+    uint32_t maxChannels;	// Number of channels in pointer array
     bmcChannel *chan;		// Pointer to the first channel element
 };
 typedef struct bmcChannels bmcChannels;
